// src/types/generated/ipc/McpInteractionResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { McpElicitationAnswer } from "./McpElicitationAnswer";
import type { McpRuntimeKey } from "./McpRuntimeKey";

export type McpInteractionResponse = { requestId: string, key: McpRuntimeKey, operationId: string, answers: Array<McpElicitationAnswer>, };